+++ legends/utils/__init__.py
@@ -3,6 +3,6 @@
 """
 
 from legends.utils.customabcs import *
-from legends.utils.relations import *
 from legends.utils.objrelations import *
+from legends.utils.relations import *
 from legends.utils.scrollframe import *